added coloration

+++ src/components/TodoListItem/todoListItem.component.ts
@@ -18,14 +18,12 @@ export class TodoListItem {
     console.log('TodoListItemComponent : list=' + JSON.stringify(this.list) + "; item=" + JSON.stringify(this.item));
   }
   
-  public deleteItem () {
+  public deleteItem (): void {
 
-    console.log('TodoListItemComponent : deleteItem : item=' + JSON.stringify(this.item) + '; list=' + JSON.stringify(this.list));
-    this.todoService.deleteTodo(this.list.uuid, this.item.uuid);
-    // if(this.todoListId != null && this.todoListId != undefined) {
-    //   console.log('deleting : ' + JSON.stringify(item));
-    //   this.todoService.deleteTodo(this.todoListId, item.uuid).subscribe().unsubscribe();
-    // }
+    if(this.list) {
+      console.log('TodoListItemComponent : deleteItem : item=' + JSON.stringify(this.item) + '; list=' + JSON.stringify(this.list));
+      this.todoService.deleteTodo(this.list.uuid, this.item.uuid);
+    }
   }
 
   /*private dataChanged($event) {
@@ -47,11 +45,21 @@ export class TodoListItem {
     }
   }*/
 
-  public editItem() {
+  public editItem(): void {
 
     this.navCtrl.push(TodoItemEditPage, {
       itemId: this.item.uuid,
       listId: this.list.uuid
     });
+  }
+
+  public completeItem(): void {
+
+    if(this.item && this.list) {
+
+      console.log('TodoListItemComponent : completeItem : item=' + JSON.stringify(this.item) + '; listId=' + JSON.stringify(this.list.uuid));
+      this.item.complete = !this.item.complete;
+      this.todoService.editTodo(this.list.uuid, this.item);
+    }
   }
 }
